+++ app/Http/Controllers/MediaController.php
@@ -6,16 +6,23 @@ use App\Http\Requests\StoreMediaRequest;
 use App\Http\Requests\UpdateMediaRequest;
 use App\Http\Resources\MediaResource;
 use App\Models\Media;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class MediaController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('role:admin')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         $media = Media::paginate(10);
 
         return MediaResource::collection($media);
@@ -26,6 +33,7 @@ class MediaController extends Controller
      */
     public function store(StoreMediaRequest $request)
     {
+        $this->authorize('create', User::class);
         $media = Media::create($request->validated());
 
         return new MediaResource($media);
@@ -36,15 +44,9 @@ class MediaController extends Controller
      */
     public function show(Media $media)
     {
-        return new MediaResource($media);
-    }
+        $this->authorize('view', $media);
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
+        return new MediaResource($media);
     }
 
     /**
@@ -52,6 +54,7 @@ class MediaController extends Controller
      */
     public function update(UpdateMediaRequest $request, Media $media)
     {
+        $this->authorize('update', $media);
         $media->update($request->validated());
 
         return new MediaResource($media);
@@ -62,6 +65,7 @@ class MediaController extends Controller
      */
     public function destroy(Media $media)
     {
+        $this->authorize('delete', $media);
         $media->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
