+++ app/Http/Controllers/MessageController.php
@@ -6,16 +6,23 @@ use App\Http\Requests\StoreMessageRequest;
 use App\Http\Requests\UpdateMessageRequest;
 use App\Http\Resources\MessageResource;
 use App\Models\Message;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class MessageController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('role:admin')->except('index', 'show');
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny', User::class);
         $messages = Message::paginate(10);
 
         return MessageResource::collection($messages);
@@ -26,6 +33,7 @@ class MessageController extends Controller
      */
     public function store(StoreMessageRequest $request)
     {
+        $this->authorize('create', User::class);
         $message = Message::create($request->validated());
 
         return new MessageResource($message);
@@ -36,6 +44,8 @@ class MessageController extends Controller
      */
     public function show(Message $message)
     {
+        $this->authorize('view', $message);
+
         return new MessageResource($message);
     }
 
@@ -44,6 +54,7 @@ class MessageController extends Controller
      */
     public function update(UpdateMessageRequest $request, Message $message)
     {
+        $this->authorize('update', $message);
         $message->update($request->validated());
 
         return new MessageResource($message);
@@ -54,6 +65,7 @@ class MessageController extends Controller
      */
     public function destroy(Message $message)
     {
+        $this->authorize('delete', $message);
         $message->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
